requirements.txt
paho-mqtt
python-dotenv
dotenv-settings-handler